Add an implementation of ceres client in command-line (#1)
* feat(cli): init crate cli and abstract commands

* feat(cli): add transaction arguments

* feat(cli): add storage implementation

* fix(cli): the tests of the util of cli

* feat(cli): store metadata in local database

* feat(cli): command info

* feat(cli): command list

* feat(cli): command call and deploy

* feat(cli): update preview in README

* feat(runtime): flush interface

* feat(cli): pretty info

+++ runtime/src/util.rs
@@ -6,7 +6,7 @@ use parity_wasm::elements::{External, Module};
 const IMPORT_MODULE_MEMORY: &str = "env";
 
 /// Parse `Vec<String>` to `Vec<RuntimeValue>`
-pub fn parse_args(selector: &str, args: &[&str], tys: Vec<u8>) -> Result<Vec<u8>> {
+pub fn parse_args(selector: &str, args: &[&str], tys: Vec<u32>) -> Result<Vec<u8>> {
     if args.len() != tys.len() {
         return Err(Error::InvalidArgumentLength);
     }
